forex_smc_system/requirements.txt
MetaTrader5>=5.0.45       # MT5 Python API (Windows only)
pandas>=2.0.0
numpy>=1.26.0
pandas-ta>=0.3.14b        # ADX, ATR, Bollinger Bands
sqlalchemy>=2.0.0
python-dateutil>=2.9.0
pydantic>=2.0.0
loguru>=0.7.0
schedule>=1.2.0
aiohttp>=3.9.0
python-telegram-bot>=20.0

# Backtesting / optimisation
optuna>=3.5.0

# Optional / dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
